HV-1245 Move the extraction of the value to validate to MetaConstraint
It simplifies ValidatorImpl and paves the way for the next steps.
